2008-09-06 11:07 UTC+0200 Viktor Szakats (harbour.01 syenar hu)    * make_vc.bat    * contrib/mtpl_vc.bat      + Added -nologo to nmake.exe command line.
   * source/vm/memvars.c
   * source/rtl/scroll.c
   * source/rtl/diskspac.c
   * source/rtl/console.c
   * source/rtl/philes.c
   * source/rtl/oldclear.c
   * source/rtl/oldbox.c
   * source/rtl/cdpapi.c
   * source/rtl/setpos.c
   * source/rtl/errorapi.c
   * source/rtl/setcolor.c
   * source/rtl/philes53.c
   * source/rtl/setkey.c
   * source/rtl/disksphb.c
   * source/rtl/shadow.c
   * source/rtl/hbhex.c
   * source/rtl/setcurs.c
   * source/rtl/dates.c
   * source/rtl/gtwvt/gtwvt.c
   * source/rtl/hbinet.c
   * source/rtl/hbffind.c
   * source/rtl/filesys.c
   * contrib/hbct/screen2.c
   * contrib/hbct/screen1.c
   * contrib/hbnf/fttext.c
     ! Fixed more MSVS 2008 -W4 warnings (about 100 of them).

   * contrib/hbnf/fttext.c
     ! Fixed C++ compile error after recently 
       switching to Harbour API from Clipper one.

+++ harbour/contrib/hbct/screen2.c
@@ -77,7 +77,7 @@ HB_FUNC( SAYDOWN )
 
       if( iRow >= 0 && iCol >= 0 && iRow <= iMaxRow && iCol <= iMaxCol )
       {
-         BYTE bColor = hb_gtGetCurrColor();
+         BYTE bColor = ( BYTE ) hb_gtGetCurrColor();
 
          if( ulLen > ( ULONG ) ( iMaxRow - iRow + 1 ) )
             ulLen = ( ULONG ) ( iMaxRow - iRow + 1 );
